pyproject.toml
[tool.poetry]
name = "diplomawork"
version = "0.1.0"
description = "DF1"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.12,<3.14"
Django = "^5.0"
djangorestframework = "^3.15"
djangorestframework-simplejwt = "^5.3.1"
drf-spectacular = "^0.27.2"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
black = "^24.8.0"
flake8 = "^7.1.0"
mypy = "^1.10.0"
pytest = "^8.3.1"
pytest-django = "^4.8.0"

[tool.black]
line-length = 88
target-version = ["py312"]

[tool.mypy]
python_version = "3.12"
ignore_missing_imports = true

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503"]
exclude = [
  ".git", "__pycache__", ".venv", "venv",
  "build", "dist",
  "*/site-packages/*",
  "**/migrations/*"
]
